+++ docs/form_qc_checker/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.5.4
+* Rebuilding for pre-processing checks update
+  
+## 1.5.3
+* Rebuilding for API call retries
+  
 ## 1.5.2
-
 * Updates `nacc-form-validator` to 0.5.3
 
 ## 1.5.1